refactor(tensor): seal `TensorKind` and `Parameter` traits and consolidate param impls
